requirements.txt
# Sentinel-AI Enterprise Workflow Engine
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
pyyaml==6.0.2

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# LLM Integration
openai==1.58.1
anthropic==0.42.0
google-generativeai==0.8.4

# Atlassian MCP Integration
httpx==0.28.1
mcp>=1.0.0

# Utilities
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Dashboard serving
aiofiles==24.1.0
websockets==14.1
python-dotenv==1.0.1
